Index Jumping
If we keep outer loop of indexes of ran and inner loop of indexes of coin then  gives a higher number of solutions that the correct answer.
The reason for this is when we create an amount array from 0...Amount, if we iterate over all the coins a solution
can be added twice. For example to create 7:

   1. When amount is 2 and the coin value is 5, it would be counted as 1 way.
   2. When amount is 5 and the coin value is 2, the number of ways become 2.

The set is either case remains 1 coin of 2 and 1 coin of 5. But our current method adds it twice.
So we create use an outer loop of coins so that a combination once used cannot be used again.