Concurrent multiple starting points for BFS
Question Link:
https://leetcode.com/problems/rotting-oranges/description/

Line 24 to 33:
Its a shortcut learnt from Striver's Video which is implemented to traverse adjacent horizontal and vertical nodes with help of for loop. The delta changes that happen while traversing are already stored in vector at line 7. We are traversing the nodes related to our current node in a clockwise fashion.

Read Comments for Better Understanding.